--- src/blocks/_pro/new-badge/style.scss
@@ -1,16 +1,0 @@
-.vk_newBadge {
-    background-color: red;
-    color: #fff;
-    border-radius: 5px;
-    font-size: 10px;
-    text-align: center;
-    padding: 0 4px;
-    word-break: break-word;
-
-    &.is-style-plain {
-        color: unset;
-        background-color: unset;
-        font-size: unset;
-        text-align: unset;
-    }
-}
