+++ src/blocks/bundle-pro.js
@@ -21,9 +21,9 @@ import * as gridColCardItemHeader from './_pro/gridcolcard-item-header';
 import * as gridColCardItemFooter from './_pro/gridcolcard-item-footer';
 import * as iconCard from './_pro/icon-card';
 import * as iconCardItem from './_pro/icon-card-item';
-import * as newBadge from './_pro/new-badge';
 import * as outer from './_pro/outer';
 import * as postList from './_pro/post-list';
+import * as postNewBadge from './_pro/post-new-badge';
 import * as selectPostList from './_pro/select-post-list';
 import * as selectPostListItem from './_pro/select-post-list-item';
 import * as step from './_pro/step';
@@ -60,9 +60,9 @@ const vkblocksPro = [
 	gridColCardItemHeader,
 	iconCard,
 	iconCardItem,
-	newBadge,
 	outer,
 	postList,
+	postNewBadge,
 	selectPostList,
 	selectPostListItem,
 	step,
